wip(grpc): cleanup, removal of deprecated graphql
