+++ packages/client/.gitignore
@@ -30,4 +30,3 @@ torii_local_db
 # svelte
 .svelte-kit
 build
-$houdini
