+++ packages/client/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "svelte-houdini-dojo",
+	"name": "@zorg/client",
 	"version": "0.0.1",
 	"private": true,
 	"scripts": {
@@ -14,7 +14,6 @@
 		"deploy:dojo": "../tot/scripts/dev_deploy.sh run; ./scripts/cp_abis.sh ../tot"
 	},
 	"devDependencies": {
-		"@graphql-codegen/cli": "^5.0.4",
 		"@sveltejs/adapter-auto": "^3.3.1",
 		"@sveltejs/kit": "^2.17.1",
 		"@sveltejs/vite-plugin-svelte": "^3.1.2",
@@ -24,8 +23,6 @@
 		"@types/react-dom": "^19.0.3",
 		"@types/three": "^0.165.0",
 		"autoprefixer": "^10.4.20",
-		"houdini": "^1.5.0",
-		"houdini-svelte": "^1.2.64",
 		"jsdom": "^24.1.3",
 		"svelte": "^4.2.19",
 		"svelte-check": "^3.8.6",
@@ -39,22 +36,6 @@
 	"type": "module",
 	"dependencies": {
 		"@cartridge/controller": "^0.5.9",
-		"@starknet-react/chains": "^3.1.2",
-		"@starknet-react/core": "^3.7.2",
-		"@sveltejs/adapter-node": "^5.2.12",
-		"@zorg/contracts": "workspace:*",
-		"buffer": "^6.0.3",
-		"envalid": "^8.0.0",
-		"get-starknet": "^3.3.3",
-		"graphql": "^16.10.0",
-		"graphql-ws": "^5.16.2",
-		"micro-starknet": "^0.2.3",
-		"postcss": "^8.5.2",
-		"three": "^0.165.0",
-		"vite-plugin-top-level-await": "^1.5.0",
-		"vite-plugin-wasm": "^3.4.1",
-		"starknet": "6.21.0",
-		"starknetkit": "^2.6.1",
 		"@dojoengine/core": "^1.1.1",
 		"@dojoengine/create-burner": "^1.1.1",
 		"@dojoengine/react": "^1.1.1",
@@ -62,6 +43,20 @@
 		"@dojoengine/sdk": "^1.1.1",
 		"@dojoengine/state": "^1.1.1",
 		"@dojoengine/torii-client": "^1.1.1",
-		"@dojoengine/utils": "^1.1.1"
+		"@dojoengine/utils": "^1.1.1",
+		"@starknet-react/chains": "^3.1.2",
+		"@starknet-react/core": "^3.7.2",
+		"@sveltejs/adapter-node": "^5.2.12",
+		"@zorg/contracts": "workspace:*",
+		"buffer": "^6.0.3",
+		"envalid": "^8.0.0",
+		"get-starknet": "^3.3.3",
+		"micro-starknet": "^0.2.3",
+		"postcss": "^8.5.2",
+		"starknet": "6.21.0",
+		"starknetkit": "^2.6.1",
+		"three": "^0.165.0",
+		"vite-plugin-top-level-await": "^1.5.0",
+		"vite-plugin-wasm": "^3.4.1"
 	}
 }
