--- packages/client/src/lib/stores/graphqlStatus.ts
@@ -1,3 +1,0 @@
-import { writable } from "svelte/store";
-
-export const graphqlStatus = writable({ isDown: false });
