+++ packages/client/vite.config.ts
@@ -1,13 +1,12 @@
 import fs from "node:fs";
 import path from "node:path";
 import { sveltekit } from "@sveltejs/kit/vite";
-import houdini from "houdini/vite";
 import { type UserConfig, defineConfig } from "vite";
 import wasm from "vite-plugin-wasm";
 import topLevelAwait from "vite-plugin-top-level-await";
 
 const config: UserConfig = {
-	plugins: [houdini(), sveltekit(), wasm(), topLevelAwait()],
+	plugins: [sveltekit(), wasm(), topLevelAwait()],
 	build: {
 		target: "esnext",
 	},
